v0.1.2
- IMP-32: `LiteNetLibTransport` connection key — constructor injection with `DefaultConnectionKey` constant fallback.
- IMP-33: Propagate disconnect reason via `INetworkTransport.OnDisconnected` — added 6-value `DisconnectReason` enum; `Listen`/`Connect` now return `bool` to surface startup failures immediately. Client handlers gate auto-reconnect to `NetworkFailure`/`ReconnectRequested` only.
- IMP-34: 64-bit `SessionMagic` with CSPRNG generation + device-binding on cold-start reconnect.
- Aligned spectator transport connection key with the main transport.
- Aligned `Connect` API deviceId with the provider pattern; added send-site logging.

+++ Klotho/Assets/Klotho/Runtime/Core/ConnectionResult.cs
@@ -26,7 +26,7 @@ namespace xpTURN.Klotho.Core
         /// Session Magic. The SyncCompleteMessage.Magic / LateJoinAccept.Magic / ReconnectAccept (via creds) value
         /// is forwarded as-is.
         /// </summary>
-        public int SessionMagic { get; set; }
+        public long SessionMagic { get; set; }
 
         /// <summary>
         /// Clock synchronization — SharedClock epoch.
